Собранный openal-soft-1.19.1

+++ ogsr_engine/xrSound/SoundRender_Core_Processor.cpp
@@ -198,7 +198,7 @@ float CSoundRender_Core::get_occlusion_to(const Fvector& hear_pt, const Fvector&
 
 		geom_DB.ray_options(CDB::OPT_CULL);
 		geom_DB.ray_query(geom_SOM, hear_pt, dir, range);
-		u32 r_cnt = geom_DB.r_count();
+		u32 r_cnt = u32(geom_DB.r_count());
 		CDB::RESULT*	_B = geom_DB.r_begin();
 
 		if (0 != r_cnt) {
@@ -251,7 +251,7 @@ float CSoundRender_Core::get_occlusion(Fvector& P, float R, Fvector* occ)
 	if (0 != geom_SOM) {
 		geom_DB.ray_options(CDB::OPT_CULL);
 		geom_DB.ray_query(geom_SOM, base, dir, range);
-		u32 r_cnt = geom_DB.r_count();
+		u32 r_cnt = u32(geom_DB.r_count());
 		CDB::RESULT*	_B = geom_DB.r_begin();
 
 		if (0 != r_cnt) {
